Learning_rate is a (NON NULL) float

+++ src/cnn/creation.c
@@ -6,7 +6,7 @@
 
 #include "include/creation.h"
 
-Network* create_network(int max_size, int learning_rate, int dropout, int initialisation, int input_dim, int input_depth) {
+Network* create_network(int max_size, float learning_rate, int dropout, int initialisation, int input_dim, int input_depth) {
     if (dropout < 0 || dropout > 100) {
         printf("Erreur, la probabilité de dropout n'est pas respecté, elle doit être comprise entre 0 et 100\n");
     }
@@ -33,7 +33,7 @@ Network* create_network(int max_size, int learning_rate, int dropout, int initia
     return network;
 }
 
-Network* create_network_lenet5(int learning_rate, int dropout, int activation, int initialisation, int input_dim, int input_depth) {
+Network* create_network_lenet5(float learning_rate, int dropout, int activation, int initialisation, int input_dim, int input_depth) {
     Network* network = create_network(8, learning_rate, dropout, initialisation, input_dim, input_depth); 
     network->kernel[0]->activation = activation;  
     network->kernel[0]->linearisation = 0;
